+++ tests/test_program_analysis.py
@@ -1,15 +1,11 @@
 import os
 import json
-import sys
 from datetime import date
-import importlib
 import xml.etree.ElementTree as ET
 import subprocess as sp
 import ast
 import pytest
 
-from delphi.GrFN.networks import GroundedFunctionNetwork
-from delphi.visualization import visualize
 from delphi.translators.for2py import (
     preprocessor,
     translate,
@@ -124,12 +120,3 @@ def test_derived_type_pythonIR_generation(derived_types_python_IR_test):
     with open("tests/data/derived-types-03.py", "r") as f:
         python_dict = f.read()
     assert derived_types_python_IR_test == python_dict
-
-
-def test_ProgramAnalysisGraph_from_GrFN():
-    sys.path.insert(0, "tests/data/GrFN/")
-    lambdas = importlib.__import__("PETPT_torch_lambdas")
-    pgm = json.load(open("tests/data/GrFN/PETPT_numpy.json", "r"))
-    G = GroundedFunctionNetwork.from_dict(pgm, lambdas)
-    PAG = G.to_ProgramAnalysisGraph()
-    visualize(PAG, save=True, filename="petpt_pag.pdf")
